// src/app/interfaces/youtube-records-api.ts
import {Snippet} from './snippet';

export interface YoutubeRecordsApi {
  kind: string;
  etag: string;
  nextPageToken: string;
  regionCode: string;
  pageInfo: {
    totalResults: number;
    resultsPerPage: number;
  };
  items: {
    kind: string;
    etag: string;
    id: {
      kind: string;
      videoId: string;
    };
    snippet: Snippet;
  }[];
}